refactor(node): centralize pino logger
