+++ nodejs/src/utils/logger.ts
@@ -1,0 +1,5 @@
+import pino from 'pino'
+
+export const logger = pino({
+  level: process.env.LOG_LEVEL ?? 'info',
+})
